Add support for references, as discussed in #6. This commit essentially rewrites the whole `LuaValue` system. The new version works as follows: * `LuaVal` contains an address which is a pointer to some type, or it encodes 61 bits of an int or a float. * The value is encoded, and at runtime it is possible to know the type of the value to which this address points to, or if the value contains an encoded number * When a user creates a LuaValue from an integer, the integer is allocated on heap, the address is encoded with the value for integers, and then the address is stored in the LuaVal. If the integer fits in 61 bits, then the 61 bits are stored in the `LuaVal`. This ensures that small integers are not allocated on the heap.
The `Gc` type implements a garbage collector, which is very useful considering
that Lua is a garbage collected language. This design allows for recursive data
types such as tables in Lua.

`LuaVal`s can be "cloned", but in the case of tables only the `Gc` is cloned,
which means that cloning returns a new reference to the table.

There are a lot of tests which ensure that `LuaVal`s work as expected.