Change several instruction implementations. There are two types of LuaTables: * `UserTable` which is just a wrapper around HashMap * `CachingTable` which does faster lookups for strings that are part of the string constants map.
`UserFunction`s also keep track of a list of upvalues. These are values that
are in scope of a function, but are not defined in the function's scope. The VM
no longer stores `_ENV` in register 0. Instead, it creates a `CachingTable`
upvalue, which is passed to every function as an upvalue. The compiler makes
sure to always store and load global values from the 1st upvalue of a function.

`VarArg` now works very similarly to `MOVR`.